==> ro/archive4/2a-processor-allegiance.ipynb <==
import pandas as pd, numpy as np, json

base_url='http://www.cdep.ro'

nmembers=json.loads(open('data/nmembers.json','r').read())

img=json.loads(open('data/img.json','r').read())

records=json.loads(open('data/records.json','r').read())

records[90]

nmembers['Calimente Mihaita']

{i['vot'] for i in records}

votes=json.loads(open('data/votes.json','r').read())

votes[9]

def get_allegiance(allegiance,voteid,outcome,names):
    if voteid not in allegiance:
        allegiance[voteid]={}
    if outcome not in allegiance[voteid]:
        allegiance[voteid][outcome]=[]
    for name in names:
        allegiance[voteid][outcome].append(name)
    return allegiance

votes_dict={v[0]['votid']:v[0] for v in votes}
#votes_dict={v['votid']:v for v in votes}

votes_dict['388']

def senate(camera):
    if camera=='1':return 'Senatori'
    else: return 'Deputați'

import unidecode

def name_normalizer(name):
    return unidecode.unidecode(name)

len(records)

eu_allegiance={}
eu_vt={}
eu_joint_allegiance={}
eu_joint_vt={}
records_parsed=set()

for i in range(len(records[:])):
    if (i%10000==0):
        print(i)
    if i not in records_parsed:
        j=records[i]
        vote=votes_dict[j['votid']]
        ts=pd.to_datetime('-'.join(vote['time'].split(' ')[0].split('.')[::-1]))
        year=int(vote['time'].split(' ')[0][-4:])
        camera=senate(vote['camera'])
        if camera not in eu_vt:eu_vt[camera]={}
        if year not in eu_vt[camera]:eu_vt[camera][year]=[]
        if camera not in eu_allegiance:eu_allegiance[camera]={}
        if year not in eu_allegiance[camera]:eu_allegiance[camera][year]={'Name':{},'Group':{},'Party':{},'Country':{}}  
        if year not in eu_joint_vt:eu_joint_vt[year]=[]
        if year not in eu_joint_allegiance:eu_joint_allegiance[year]={'Name':{},'Group':{},'Party':{},'Country':{}}    
        if vote['descriere'] not in ["Modification de l'ordre du jour"]:
                #name=' '.join(j['name'].split(' ')[::-1])
                name=name_normalizer(j['name'])
                if name not in nmembers:
                    print(name)
                else:
                    member=nmembers[name]
                    dummy={}
                    outcome=j['vot']
                    dummy['Vote']=j['votid']
                    dummy['Name']=[member['Name']['full']]
                    dummy['Outcome']=outcome
                    dummy['Year']=year
                    dummy['Group']=[]
                    dummy['Party']=[]
                    dummy['Country']=[]
                    for k in member['Parties']:
                        if pd.to_datetime(k['start'])<ts<pd.to_datetime(k['end']):
                            dummy['Party'].append(k['party'])
                    for k in member['Groups']:
                        if pd.to_datetime(k['start'])<ts<pd.to_datetime(k['end']):
                            dummy['Group'].append(k['group'])
                    for k in member['Countries']:
                        if pd.to_datetime(k['start'])<ts<pd.to_datetime(k['end']):
                            dummy['Country'].append(k['country'])
                    eu_vt[camera][year].append(dummy)
                    eu_joint_vt[year].append(dummy)
                    for allegiance_type in eu_allegiance[camera][year].keys():
                        eu_allegiance[camera][year][allegiance_type]=\
                            get_allegiance(eu_allegiance[camera][year][allegiance_type],j['votid'],
                                       outcome,dummy[allegiance_type])
                        eu_joint_allegiance[year][allegiance_type]=\
                            get_allegiance(eu_joint_allegiance[year][allegiance_type],j['votid'],
                                       outcome,dummy[allegiance_type])
        records_parsed.add(i)

eu_allegiance['Joint']=eu_joint_allegiance
eu_vt['Joint']=eu_joint_vt

open('data/eu_allegiance.json','w').write(json.dumps(eu_allegiance))

open('data/eu_vt.json','w').write(json.dumps(eu_vt))
