==> ro/archive4/3d-plotter-party.ipynb <==
import pandas as pd, numpy as np, json, os
import matplotlib.pyplot as plt
%matplotlib inline

plt.style.use('seaborn-whitegrid')

#!pip install d3IpyPlus
from d3IpyPlus import ScatterPlot, LinePlot, BarPlot, StackedArea

#!pip install pyecharts

m=json.loads(open('ignore/member_timelines.json','r').read())
members=json.loads(open('ignore/member_simple.json','r').read())

d=pd.DataFrame(members['deaths'],index=['v']).T

d['v']=' †'

s=pd.DataFrame(members['s2'],index=['s']).T
s=s.join(d)
s['v']=s['v'].fillna('')
s['s']=s['s']+s['v']
s=s[['s']]
s['w']=s['s'].str.split(' ').str[0].str[0]+s['s'].str.split(' ').str[-1].str[0]

szd_color='#E91C62'

pcolors=json.loads(open('json/pcolors.json','r').read())

g=pd.DataFrame(members['megye'],index=['Megye ']).T
g.head()

data={}
attention={}
for k,i in enumerate(m):
    if k%1000==0: print(k/len(m)*100,'%')
    if i['Típus']=='Párt':
        if i['Csoportok'] not in data:data[i['Csoportok']]={'start':pd.to_datetime('2030'),'end':pd.to_datetime('1980')}
        start=pd.to_datetime(i['start'])
        end=pd.to_datetime(i['end'])
        data[i['Csoportok']]['start']=min(data[i['Csoportok']]['start'],start)
        data[i['Csoportok']]['end']=max(data[i['Csoportok']]['end'],end)

part_timelines=pd.DataFrame(data).T
part_timelines.index.name='id'

part_timelines['start']=part_timelines['start'].astype(str).str[:10]
part_timelines['end']=part_timelines['end'].astype(str).str[:10]

p=part_timelines.reset_index()[['id']]
p['Color']=p['id'].str.split(' ').str[1].str.strip()
p['Part']=p['id'].str.split('-').str[0].str.strip()

part_timelines=part_timelines.join(p.set_index('id').drop_duplicates())

logos={'🕊️':'ALDE', '🌹':'PSD', '📐':'PNL', '🔶':'PD', '⚛️':'PP-DD', '👤':'Independent', '⚪️':'FSN',
       '✳️':'PNTCD', '🔱':'PRM', '🌷':'RMDSZ', '🍀':'ECO', '🍏':'PMP',
       '🔷':'USR', '🦅':'UNPR', '🔘':'Mas'}

part_timelines['logo']=part_timelines['Part'].str.split(' ').str[0]

part_timelines['group']=[logos[i] for i in part_timelines['logo'].values]
part_timelines['label']=[i+' '+logos[i] for i in part_timelines['logo'].values]

pcolors=json.loads(open('json/pcolors.json','r').read())

part_timelines=part_timelines.join(pd.DataFrame(pcolors,index=['cvalue']).T,on='group').fillna('grey')

part_timelines

part_timelines_list=list(part_timelines.reset_index().T.to_dict().values())

open('json/part_timelines.json','w').write(json.dumps(part_timelines_list))

Resolve for non-activity


legs={'2016':['2016-12-21','2019-07-01'],
'2012':['2012-12-20','2016-12-21'],
'2008':['2008-12-19','2012-12-20'],
'2004':['2004-12-19','2008-12-13'],
'2000':['2000-12-15','2004-11-30'],
'1996':['1996-11-27','2000-11-30'],
'1992':['1992-10-28','1996-11-22'],
'1990':['1990-06-19','1992-10-16']}

switch_months=set([str(pd.to_datetime(j))[:7] for i in legs for j in legs[i]])

from calendar import monthrange

data2={}
parties={}
for k,i in enumerate(m):
    if k%1000==0: print(k/len(m)*100,'%')
    if i['Típus']=='Párt':
        name=i['Name']
        part=i['Csoportok']
        if part not in parties:parties[part]=[]
        parties[part].append(name)
        if part not in data2:data2[part]={}
        for t in pd.date_range(i['start'],i['end']):
            st=str(t)[:10]
            if st not in data2[part]:data2[part][st]=0
            data2[part][st]+=1

df=pd.DataFrame(data2)

df=df.stack().reset_index()
df.columns=['x','id','y']

df=df.set_index(['id','x'])

dqs=[]
for d in df.index.get_level_values(0).unique():
    dq=df.loc[d]
    dq.index=pd.to_datetime(dq.index)
    dq=dq.sort_index()
    dq=dq.resample('1M').sum()
    dq['id']=d
    dq['month_length']=[monthrange(t.year,t.month)[1] for t in dq.index]
    dqs.append(dq)
df=pd.concat(dqs).reset_index()

df=df.sort_values(by='id')

df['y']=df['y']/df['month_length']

#df=df.groupby(['id','x']).mean().rolling(12).mean().reset_index().dropna()

df2=df.join(part_timelines,on='id')

df2['x']=df2['x'].astype(str).str[:7]

df2=df2[df2['y']>0]
values=list(df2.T.to_dict().values())
open('json/part_timelines2.json','w').write(json.dumps(values))
